fix up golanglint ci, bk pipeline, linter config; moves schemagen from root -> pkg
