+++ .buildkite/pipeline.yaml
@@ -13,7 +13,7 @@ steps:
     key: "generate"
     plugins:
       - docker#v5.10.0:
-          image: "ghcr.io/datumforge/base-ci-image:28-5226be2e"
+          image: "ghcr.io/datumforge/base-ci-image:v1.1.3"
           command: ["task", "ci"]
           environment: 
             - "GOTOOLCHAIN=auto"
@@ -21,7 +21,7 @@ steps:
     key: "generate_config"
     plugins:
       - docker#v5.10.0:
-          image: "ghcr.io/datumforge/base-ci-image:28-5226be2e"
+          image: "ghcr.io/datumforge/base-ci-image:v1.1.3"
           command: ["task", "config:ci"]
           environment: 
             - "GOTOOLCHAIN=auto"
@@ -34,8 +34,8 @@ steps:
     key: "lint"
     plugins:
       - docker#v5.10.0:
-          image: "registry.hub.docker.com/golangci/golangci-lint:v1.55.2"
-          command: ["golangci-lint", "run", "-v", "--timeout", "10m"]
+          image: "registry.hub.docker.com/golangci/golangci-lint:v1.57.1"
+          command: ["golangci-lint", "run", "-v", "--timeout", "10m", "--config", ".golangci.yaml"]
           environment: 
             - "GOTOOLCHAIN=auto"
 
@@ -190,7 +190,7 @@ steps:
   steps:
   - label: ":docker: docker pr build"
     key: "docker-pr-build"
-    if: build.branch != "main"
+    if: build.branch != "main" && build.tag == null
     commands: |
       #!/bin/bash
       ls
@@ -263,3 +263,57 @@ steps:
           ignore-unfixed: true
           security-checks: config,secret,vuln
           skip-files: 'cosign.key,Dockerfile.dev'
+  - label: ":docker: docker build and publish"
+    key: "docker-build"
+    if: build.tag != null
+    commands: |
+      #!/bin/bash
+      ls
+    plugins:
+      - docker-login#v3.0.0:
+          username: datum-bender
+          password-env: SECRET_GHCR_PUBLISH_TOKEN
+          server: ghcr.io
+      - datumforge/docker-metadata#v1.0.0:
+          images:
+          - "${IMAGE_REPO}"
+          extra_tags:
+          - "${BUILDKITE_TAG}"
+      - datumforge/container-build#v1.1.0:
+          dockerfile: docker/Dockerfile
+          push: true
+          build-args:
+            - NAME=${APP_NAME}
+      - equinixmetal-buildkite/trivy#v1.18.5:
+          severity: CRITICAL,HIGH
+          ignore-unfixed: true
+          security-checks: config,secret,vuln
+          skip-files: 'cosign.key,Dockerfile.dev'
+  - label: ":docker: docker build and publish all in one"
+    key: "docker-build-aio"
+    if: build.tag != null
+    commands: |
+      #!/bin/bash
+    plugins:
+      - docker-login#v3.0.0:
+          username: _json_key_base64
+          password-env: SECRET_GCR_PUBLISH_TOKEN
+          server: us-central1-docker.pkg.dev
+      - datumforge/docker-metadata#v1.0.0:
+          tag_prefix: "aio-"
+          debug: true
+          images:
+          - "${GCR_REPO}"
+          extra_tags:
+          - "${BUILDKITE_TAG}"
+      - datumforge/container-build#v1.1.0:
+          dockerfile: docker/all-in-one/Dockerfile.all-in-one
+          push: true
+          build-args:
+            - NAME=${APP_NAME}
+      - equinixmetal-buildkite/trivy#v1.18.5:
+          severity: CRITICAL,HIGH
+          ignore-unfixed: true
+          security-checks: config,secret,vuln
+          skip-files: 'cosign.key,Dockerfile.dev'
+
