+++ .golangci.yaml
@@ -1,3 +1,7 @@
+run:
+  timeout: 10m
+  allow-serial-runners: true
+
 linters-settings:
   goimports:
     local-prefixes: github.com/datumforge/datum
@@ -8,62 +12,40 @@ linters-settings:
   revive:
     ignore-generated-header: true
 
-run:
-  skip-files:
-    - internal/entclean/main.go
-    - pkg/testutils
-    - .buildkite/*
-    - .github/*
-    - docker/*
-    - schema/*
-    - pkg/utils/totp/testing/*
-
 linters:
   enable:
-    # default linters
-    - errcheck
-    - gosimple
-    - govet
-    - ineffassign
-    - staticcheck
-    - typecheck
-    - unused
-    - gosec
     - bodyclose
-    - noctx
+    - errcheck
     - gocritic
     - gocyclo
     - goerr113
     - gofmt
     - goimports
     - gomnd
+    - gosimple
     - govet
+    - gosec
+    - ineffassign
     - misspell
+    - noctx
     - revive
+    - staticcheck
     - stylecheck
+    - typecheck
+    - unused
     - whitespace
     - wsl
 issues:
-  exclude:
-    # Default excludes from `golangci-lint run --help` with EXC0002 removed
-    # EXC0001 errcheck: Almost all programs ignore errors on these functions and in most cases it's ok
-    - Error return value of .((os\.)?std(out|err)\..*|.*Close|.*Flush|os\.Remove(All)?|.*print(f|ln)?|os\.(Un)?Setenv). is not checked
-    # EXC0002 golint: Annoying issue about not having a comment. The rare codebase has such comments
-    # - (comment on exported (method|function|type|const)|should have( a package)? comment|comment should be of the form)
-    # EXC0003 golint: False positive when tests are defined in package 'test'
-    - func name will be used as test\.Test.* by other packages, and that stutters; consider calling this
-    # EXC0004 govet: Common false positives
-    - (possible misuse of unsafe.Pointer|should have signature)
-    # EXC0005 staticcheck: Developers tend to write in C-style with an explicit 'break' in a 'switch', so it's ok to ignore
-    - ineffective break statement. Did you mean to break out of the outer loop
-    # EXC0006 gosec: Too many false-positives on 'unsafe' usage
-    - Use of unsafe calls should be audited
-    # EXC0007 gosec: Too many false-positives for parametrized shell calls
-    - Subprocess launch(ed with variable|ing should be audited)
-    # EXC0008 gosec: Duplicated errcheck checks
-    - (G104|G307)
-    # EXC0009 gosec: Too many issues in popular repos
-    - (Expect directory permissions to be 0750 or less|Expect file permissions to be 0600 or less)
-    # EXC0010 gosec: False positive is triggered by 'src, err := ioutil.ReadFile(filename)'
-    - Potential file inclusion via variable
-  exclude-use-default: false
+  fix: true
+  exclude-use-default: true
+  exclude-dirs:
+    - pkg/testutils/*
+    - .buildkite/*
+    - .github/*
+    - docker/*
+    - pkg/utils/totp/testing/*
+    - internal/ent/generated/*
+  exclude-files:
+    - internal/graphapi/gen_server.go
+    - internal/graphapi/
+    - internal/datumclient/graphclient.go
