+++ jsonschema/schema_generator.go
@@ -6,11 +6,12 @@ import (
 	"os"
 	"reflect"
 
-	"github.com/datumforge/datum/config"
-	"github.com/datumforge/datum/jsonschema/envparse"
 	"github.com/invopop/jsonschema"
 	"github.com/invopop/yaml"
 	"github.com/mcuadros/go-defaults"
+
+	"github.com/datumforge/datum/config"
+	"github.com/datumforge/datum/jsonschema/envparse"
 )
 
 // const values used for the schema generator
@@ -124,6 +125,7 @@ func generateSchema(c schemaConfig, structure interface{}) error {
 	// generate the environment variables from the config
 	envSchema := ""
 	configMapSchema := "\n"
+
 	for _, k := range out {
 		defaultVal := k.Tags.Get(defaultTag)
