+++ pkg/utils/oas/base.yaml
@@ -1,7 +1,7 @@
 openapi: 3.1.0
 info:
-    title: Datum OpenAPI 3.1.0 Specifications
-    description: Documentation for Datum's API services
+    title: Datum Example OpenAPI 3.1.0 Specifications
+    description: Documentation for Datum's Example API services
     termsOfService: https://datum.net/terms
     contact:
         name: Datum Support
@@ -12,11 +12,11 @@ info:
         url: https://www.apache.org/licenses/LICENSE-2.0.html
     version: 1.0.1
 servers:
-    - url: http://localhost:17608/v1
-    - url: https://api.datum.net/v1
+    - url: https://example.datum.net/v1
 paths:
     /ready:
         get:
+            operationId: ready
             security:
                 - {}
             summary: Status Check
@@ -34,4 +34,13 @@ paths:
                                     redis:
                                         type: string
                                     sqlite_db_primary:
+                                        type: string
+                "400":
+                    description: The server is not ready
+                    content:
+                        application/json:
+                            schema:
+                                type: object
+                                properties:
+                                    error:
                                         type: string
